requirements.txt
discord.py==2.5.0
python-dotenv==1.0.1
python-telegram-bot==21.11.1